requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
dj-database-url==1.2.0
dj-rest-auth==3.0.0
Django==4.2
django-allauth==0.52.0
django-cors-headers==3.14.0
django-simpleui==2023.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
mysqlclient==2.1.1
oauthlib==3.2.2
openai==0.27.5
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2022.7.1
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.4.1
sqlparse==0.4.3
stats==0.1.2a0
tiktoken==0.3.3
tqdm==4.65.0
uritemplate==4.1.1
urllib3==1.26.15
yarl==1.9.2
